+++ plugins/catalog-react/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @backstage/plugin-catalog-react
 
+## 0.2.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/integration@0.5.8
+  - @backstage/core-components@0.1.5
+  - @backstage/catalog-model@0.9.0
+  - @backstage/catalog-client@0.3.16
+
 ## 0.2.5
 
 ### Patch Changes