+++ plugins/search-backend/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @backstage/plugin-search-backend
 
+## 0.2.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.8.5
+  - @backstage/plugin-search-backend-node@0.3.0
+
 ## 0.2.1
 
 ### Patch Changes